+++ lib/kaffy/scheduler/supervisor.ex
@@ -3,7 +3,7 @@ defmodule Kaffy.Scheduler.Supervisor do
 
   def start_link(args) do
     result = DynamicSupervisor.start_link(__MODULE__, args, name: KaffyTaskSupervisor)
-    tasks = Kaffy.ResourceAdmin.collect_tasks()
+    tasks = Kaffy.Tasks.collect_tasks()
 
     for task <- tasks do
       DynamicSupervisor.start_child(KaffyTaskSupervisor, {Kaffy.Scheduler.Task, task})
